requirements.txt
wheel
python-dotenv
boto3
git+https://github.com/2noise/ChatTTS.git
fastapi
pydantic
soundfile
torch
torchaudio
uvicorn